requirements.txt
fastapi
uvicorn
python-multipart
pymongo
motor
python-dotenv
pdfminer.six
PyPDF2
PyMuPDF
flake8
mypy